fix(types): zero mypy errors across PR + clean up core type annotations
## Tests
Replaced `X = None` fallback pattern in try/except ImportError blocks with
`pytest.skip(..., allow_module_level=True)`. mypy cannot reconcile
`type[X] = None` assignments; module-level skip is the canonical idiom.

Files fixed:
- tests/core/test_compliance_modules.py
- tests/core/test_ml_modules.py
- tests/core/test_physics_modules.py
- tests/core/test_security_modules.py
- tests/execution/test_risk_core.py
- tests/analytics/test_irreversibility.py (removed broken IGSStreamer block)

## Core (pre-existing mypy errors surfaced by new test imports)

**core/kuramoto/delayed.py**
- `history_fn: callable | None` → `Callable[[float], NDArray[np.float64]] | None`
- Cast `tau` scalar path to float with `# type: ignore[arg-type]` for numpy scalar.

**core/kuramoto/ricci_flow_engine.py**
- Optional Ollivier/Forman import fallback annotated with `Any` for `_lite_ricci`.
- `_to_light_graph` local `lite` variable typed as `Any` to reconcile the
  LightGraph / _SimpleUndirectedGraph fallback branch.

**core/physics/higher_order_kuramoto.py**
- `omega` / `theta0` Optional parameters resolved into typed local arrays
  (`omega_arr`, `theta0_arr`) via `np.asarray(..., dtype=np.float64)` so the
  downstream RK4 call receives a concrete `NDArray[np.float64]`.
- `corr` / return path likewise normalised through `np.asarray`.

**core/physics/uncertainty.py**
- Cast `heisenberg_uncertainty` result to `float` (it may return numpy scalar).
- `violation_factor` and `is_valid` explicitly cast to Python built-ins.

**core/physics/forman_ricci.py** / **liquidity_coupling.py** / **wave_propagation.py**
- Narrow numpy return types via `np.asarray(..., dtype=np.float64)` and
  `.astype(np.float64)` where the upstream call produced a union type.

## Result

- 0 mypy errors across 37 PR source files.
- 0 ruff warnings.
- `black --check` passes.
- 771 tests pass, 1 skipped (behavioural profiler optional dep).

+++ core/physics/wave_propagation.py
@@ -175,7 +175,7 @@ class GraphDiffusionEngine:
         - Second smallest = algebraic connectivity (Fiedler value)
         """
         eigenvalues = np.linalg.eigvalsh(L)
-        return eigenvalues
+        return eigenvalues.astype(np.float64)
 
 
 __all__ = ["GraphDiffusionEngine"]
